Move configSalt validation from deserialization. Added platform based line ending. Moved LogHelper into EvaluateLogger.
