+++ src/main/java/com/configcat/Utils.java
@@ -3,28 +3,14 @@ package com.configcat;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 
-import java.text.DecimalFormat;
-import java.text.DecimalFormatSymbols;
-import java.util.Locale;
-
-
 final class Utils {
     private Utils() { /* prevent from instantiation*/ }
 
     static final Gson gson = new GsonBuilder().disableHtmlEscaping().create();
 
-    public static DecimalFormat getDecimalFormat() {
-        DecimalFormat decimalFormat = new DecimalFormat("0.#####");
-        decimalFormat.setDecimalFormatSymbols(DecimalFormatSymbols.getInstance(Locale.UK));
-        return decimalFormat;
-    }
-
     public static Config deserializeConfig(String json) {
         Config config = Utils.gson.fromJson(json, Config.class);
         String salt = config.getPreferences().getSalt();
-        if (salt == null || salt.isEmpty()) {
-            throw new IllegalArgumentException("Config JSON salt is missing.");
-        }
         Segment[] segments = config.getSegments();
         if (segments == null) {
             segments = new Segment[]{};
